ai-service/requirements.txt
# ODAN AI Service Dependencies

# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# HTTP Client
httpx==0.26.0
aiohttp==3.9.1

# AI/ML
transformers==4.37.0
torch==2.1.2
pillow==10.2.0
numpy==1.26.3

# HuggingFace Hub
huggingface-hub==0.20.2

# Utilities
python-dotenv==1.0.0
pydantic==2.5.3
pydantic-settings==2.1.0

# For local model caching
accelerate==0.26.1
